LayoutTests/platform/chromium/fast/loader/page-dismissal-modal-dialogs-expected.txt

frame "<!--framePath //<!--frame0-->-->" - has 1 onunload handler(s)
CONSOLE MESSAGE: line 1: Blocked alert('mainFrame beforeunload PASS') during beforeunload.
CONSOLE MESSAGE: line 1: Blocked confirm('mainFrame beforeunload PASS') during beforeunload.
CONSOLE MESSAGE: line 1: Blocked prompt('mainFrame beforeunload PASS') during beforeunload.
CONSOLE MESSAGE: line 1: Blocked alert('iFrame click PASS') during beforeunload.
CONSOLE MESSAGE: line 1: Blocked confirm('iFrame click PASS') during beforeunload.
CONSOLE MESSAGE: line 1: Blocked prompt('iFrame click PASS') during beforeunload.
CONSOLE MESSAGE: line 1: Blocked alert('iFrame beforeunload PASS') during beforeunload.
CONSOLE MESSAGE: line 1: Blocked confirm('iFrame beforeunload PASS') during beforeunload.
CONSOLE MESSAGE: line 1: Blocked prompt('iFrame beforeunload PASS') during beforeunload.
CONSOLE MESSAGE: line 1: Blocked alert('mainFrame click PASS') during beforeunload.
CONSOLE MESSAGE: line 1: Blocked confirm('mainFrame click PASS') during beforeunload.
CONSOLE MESSAGE: line 1: Blocked prompt('mainFrame click PASS') during beforeunload.
CONSOLE MESSAGE: line 1: Blocked alert('mainFrame pagehide PASS') during pagehide.
CONSOLE MESSAGE: line 1: Blocked confirm('mainFrame pagehide PASS') during pagehide.
CONSOLE MESSAGE: line 1: Blocked prompt('mainFrame pagehide PASS') during pagehide.
CONSOLE MESSAGE: line 1: Blocked alert('iFrame click PASS') during pagehide.
CONSOLE MESSAGE: line 1: Blocked confirm('iFrame click PASS') during pagehide.
CONSOLE MESSAGE: line 1: Blocked prompt('iFrame click PASS') during pagehide.
CONSOLE MESSAGE: line 1: Blocked alert('mainFrame unload PASS') during unload.
CONSOLE MESSAGE: line 1: Blocked confirm('mainFrame unload PASS') during unload.
CONSOLE MESSAGE: line 1: Blocked prompt('mainFrame unload PASS') during unload.
CONSOLE MESSAGE: line 1: Blocked alert('iFrame click PASS') during unload.
CONSOLE MESSAGE: line 1: Blocked confirm('iFrame click PASS') during unload.
CONSOLE MESSAGE: line 1: Blocked prompt('iFrame click PASS') during unload.
CONSOLE MESSAGE: line 1: Blocked alert('iFrame pagehide PASS') during pagehide.
CONSOLE MESSAGE: line 1: Blocked confirm('iFrame pagehide PASS') during pagehide.
CONSOLE MESSAGE: line 1: Blocked prompt('iFrame pagehide PASS') during pagehide.
CONSOLE MESSAGE: line 1: Blocked alert('iFrame unload PASS') during unload.
CONSOLE MESSAGE: line 1: Blocked confirm('iFrame unload PASS') during unload.
CONSOLE MESSAGE: line 1: Blocked prompt('iFrame unload PASS') during unload.
PASS
